requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
openai>=1.30,<2.0
numpy>=1.24,<3.0
pandas>=2.0,<3.0
passlib[bcrypt]>=1.7,<2.0
python-jose>=3.3,<4.0
opencv-python-headless>=4.9,<5.0
ultralytics>=8.2,<9.0
scikit-learn>=1.4,<2.0
